move system profile graphs to compressed air directory and apply styles on during printing
