+++ src/app/compressed-air-assessment/compressed-air-report/compressed-air-report.module.ts
@@ -16,7 +16,7 @@ import { FacilityInfoSummaryModule } from '../../shared/facility-info-summary/fa
 import { SharedCompressorPipesModule } from '../shared-compressor-pipes/shared-compressor-pipes.module';
 import { PrintOptionsMenuModule } from '../../shared/print-options-menu/print-options-menu.module';
 import { CompressorSummaryTableModule } from '../compressor-summary-table/compressor-summary-table.module';
-import { SystemProfileGraphsModule } from '../../shared/system-profile-graphs/system-profile-graphs.module';
+import { SystemProfileGraphsModule } from '../system-profile-graphs/system-profile-graphs.module';
 import { InventoryPerformanceProfileModule } from '../../shared/inventory-performance-profile/inventory-performance-profile.module';
 import { PerformanceProfilesComponent } from './performance-profiles/performance-profiles.component';
