+++ src/app/compressed-air-assessment/system-profile-graphs/system-profile-graphs.component.html
@@ -1,0 +1,21 @@
+<div class="print-container">
+    <div class="print-avoid-break">
+        <label class="group-label bg-white ml-2">{{labelName}} System Capacity</label>
+        <div class="w-100" #systemCapacityGraph id="systemCapacityGraph"></div>
+    </div>
+    <hr>
+    <div class="print-avoid-break">
+        <label class="group-label bg-white ml-2">{{labelName}} Compressor Airflow Capacity</label>
+        <div class="w-100" #compressorCapacityGraph id="compressorCapacityGraph"></div>
+    </div>
+    <hr>
+    <div class="print-avoid-break">
+        <label class="group-label bg-white ml-2">{{labelName}} System Power</label>
+        <div class="w-100" #systemPowerGraph id="systemPowerGraph"></div>
+    </div>
+    <hr>
+    <div class="print-avoid-break">
+        <label class="group-label bg-white ml-2">{{labelName}} Compressor Power</label>
+        <div class="w-100" #compressorPowerGraph id="compressorPowerGraph"></div>
+    </div>
+</div>
