--- src/app/shared/system-profile-graphs/system-profile-graphs.component.html
@@ -1,21 +1,0 @@
-<div class="print-container">
-    <div class="print-avoid-break">
-        <label class="group-label bg-white pl-2">{{labelName}} System Capacity</label>
-        <div #systemCapacityGraph id="systemCapacityGraph"></div>
-    </div>
-    <hr>
-    <div class="print-avoid-break">
-        <label class="group-label bg-white pl-2">{{labelName}} Compressor Airflow Capacity</label>
-        <div #compressorCapacityGraph id="compressorCapacityGraph"></div>
-    </div>
-    <hr>
-    <div class="print-avoid-break">
-        <label class="group-label bg-white pl-2">{{labelName}} System Power</label>
-        <div #systemPowerGraph id="systemPowerGraph"></div>
-    </div>
-    <hr>
-    <div class="print-avoid-break">
-        <label class="group-label bg-white pl-2">{{labelName}} Compressor Power</label>
-        <div #compressorPowerGraph id="compressorPowerGraph"></div>
-    </div>
-</div>
